requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.0.6
django-autoslug==1.9.8
django-bootstrap-v5==1.0.11
django-ckeditor==6.4.2
django-js-asset==2.0.0
Pillow==9.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
